requirements.txt
Flask
Flask-SocketIO
google-cloud-vision
opencv-python
numpy
requests
Pillow
beautifulsoup4
selenium
